chore: rename shared-lib package
