+++ .prettierrc
@@ -3,7 +3,7 @@
     "@trivago/prettier-plugin-sort-imports",
     "prettier-plugin-tailwindcss"
   ],
-  "importOrder": ["^@/(.*)$", "^[./]"],
+  "importOrder": ["^@vtuber-stamp-rally/(.*)$", "^@/(.*)$", "^[./]"],
   "importOrderSeparation": true,
   "importOrderSortSpecifiers": true,
   "overrides": [
