+++ functions/modules/get-private-key/tsconfig.json
@@ -4,6 +4,7 @@
     "strict": true,
     "module": "NodeNext",
     "target": "ESNext",
+    "allowImportingTsExtensions": true,
     "moduleResolution": "NodeNext"
   },
   "include": ["src"]
