+++ scripts/delete-users-test.tsx
@@ -1,5 +1,5 @@
+import { StandistDocument } from "@vtuber-stamp-rally/shared-lib/models/Standist.ts";
 import * as sdk from "node-appwrite";
-import { StandistDocument } from "shared-lib";
 
 import { getEnv } from "./shared.js";
 import { deleteUserMedia } from "./upload-user-medias.js";
