+++ scripts/tsconfig.json
@@ -2,6 +2,8 @@
   "compilerOptions": {
     "module": "NodeNext",
     "target": "ESNext",
-    "moduleResolution": "NodeNext"
+    "noEmit": true,
+    "moduleResolution": "NodeNext",
+    "allowImportingTsExtensions": true
   }
 }
