--- shared-lib/src/index.ts
@@ -1,8 +1,0 @@
-export * from "./base64";
-export * from "./models";
-export * from "./crypto";
-export * from "./signatures";
-export * from "./types";
-export type { GetPrivateKeyFunctionResponse } from "./functions/getPrivateKey";
-export type { RegisterContestParticipantFunctionResponse } from "./functions/registerContestParticipant";
-export type { SubmitRallyFunctionResponse } from "./functions/submitRally";
