--- shared-lib/src/models/index.ts
@@ -1,6 +1,0 @@
-export * from "./ContestParticipant";
-export * from "./ContestParticipation";
-export * from "./KeyValue";
-export * from "./Standist";
-export * from "./Stamp";
-export * from "./Submission";
