+++ shared-lib/tsconfig.json
@@ -7,6 +7,7 @@
     "moduleResolution": "Bundler",
     "lib": ["ESNext", "DOM"],
     "esModuleInterop": true,
+    "noEmit": true,
     "strict": true
   },
   "include": ["src/**/*.ts"]
