+++ web/package.json
@@ -7,7 +7,7 @@
     "dev": "vite",
     "build": "tsc -b && vite build",
     "lint": "eslint . --report-unused-disable-directives --max-warnings 0",
-    "check-types": "tsc --noEmit",
+    "check-types": "tsc",
     "preview": "vite preview",
     "test": "LANG=C vitest",
     "test:ui": "LANG=C vitest --ui",
@@ -50,7 +50,7 @@
     "react-hook-form": "^7.54.2",
     "react-i18next": "^15.4.0",
     "react-qr-code": "^2.0.15",
-    "shared-lib": "workspace:*",
+    "@vtuber-stamp-rally/shared-lib": "workspace:*",
     "vaul": "^1.1.2",
     "workbox-cacheable-response": "^7.3.0",
     "workbox-core": "^7.3.0",
