+++ web/src/components/routes/rallyists/contest/SuccessPage.tsx
@@ -1,8 +1,8 @@
 import { useConfetti } from "@stevent-team/react-party";
+import { ContestParticipant } from "@vtuber-stamp-rally/shared-lib/models/ContestParticipant.ts";
 import { Check, Gift } from "lucide-react";
 import { LegacyRef, useEffect, useState } from "react";
 import { useTranslation } from "react-i18next";
-import { ContestParticipant } from "shared-lib";
 
 import { ButtonLink } from "@/components/controls/ButtonLink";
 import { Header } from "@/components/layout/Header";
