+++ web/src/components/routes/staff/StampLinks/StampLinksArtistsList.tsx
@@ -1,5 +1,5 @@
+import { Standist } from "@vtuber-stamp-rally/shared-lib/models/Standist.ts";
 import { ArrowUpRightFromSquare } from "lucide-react";
-import { Standist } from "shared-lib";
 
 import { ButtonLink } from "@/components/controls/ButtonLink";
 import { Header } from "@/components/layout/Header.tsx";
