+++ web/src/components/scan/StampDetails.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from "@storybook/react";
-import { Standist } from "shared-lib";
+import { Standist } from "@vtuber-stamp-rally/shared-lib/models/Standist.ts";
 
 import { NotSubmittedStamp, OtherStamps } from "@/stubs/Stamp.ts";
 import { StandistsFromAppwrite } from "@/stubs/Standists.ts";
