+++ web/src/components/scan/StampDetails.tsx
@@ -1,3 +1,5 @@
+import type { Stamp } from "@vtuber-stamp-rally/shared-lib/models/Stamp.ts";
+import type { Standist } from "@vtuber-stamp-rally/shared-lib/models/Standist.ts";
 import {
   CalendarCogIcon,
   Printer,
@@ -8,7 +10,6 @@ import {
 } from "lucide-react";
 import { FC, PropsWithChildren } from "react";
 import { useTranslation } from "react-i18next";
-import type { Stamp, Standist } from "shared-lib";
 
 type StampDetailsProps = {
   stamp: Stamp;
