+++ web/src/lib/db.ts
@@ -1,8 +1,20 @@
+import {
+  ContestParticipation,
+  contestParticipationIndexes,
+} from "@vtuber-stamp-rally/shared-lib/models/ContestParticipation.ts";
+import {
+  StampWithId,
+  stampIndexes,
+} from "@vtuber-stamp-rally/shared-lib/models/Stamp.ts";
+import {
+  Standist,
+  standistIndexes,
+} from "@vtuber-stamp-rally/shared-lib/models/Standist.ts";
+import {
+  SubmissionWithId,
+  submissionIndexes,
+} from "@vtuber-stamp-rally/shared-lib/models/Submission.ts";
 import Dexie, { type EntityTable } from "dexie";
-import { ContestParticipation, contestParticipationIndexes } from "shared-lib";
-import { StampWithId, stampIndexes } from "shared-lib";
-import { Standist, standistIndexes } from "shared-lib";
-import { SubmissionWithId, submissionIndexes } from "shared-lib";
 
 export const db = new Dexie("StampRally") as Dexie & {
   stamps: EntityTable<StampWithId, "id">;
