+++ web/src/lib/hooks/contest/useContestWinners.ts
@@ -1,6 +1,6 @@
 import { useQuery } from "@tanstack/react-query";
+import type { ContestParticipant } from "@vtuber-stamp-rally/shared-lib/models/ContestParticipant.ts";
 import { Query } from "appwrite";
-import type { ContestParticipant } from "shared-lib";
 
 import { QUERY_KEYS } from "@/lib/QueryKeys";
 import { databases } from "@/lib/appwrite";
