+++ web/src/lib/hooks/contest/useFollowParticipation.ts
@@ -1,8 +1,8 @@
 import { captureException } from "@sentry/react";
 import { skipToken, useQuery } from "@tanstack/react-query";
+import { ContestParticipant } from "@vtuber-stamp-rally/shared-lib/models/ContestParticipant.ts";
 import { Query, RealtimeResponseEvent } from "appwrite";
 import { useCallback, useEffect } from "react";
-import { ContestParticipant } from "shared-lib";
 
 import { QUERY_KEYS } from "@/lib/QueryKeys";
 import { client, databases } from "@/lib/appwrite";
