+++ web/src/lib/hooks/contest/useRegisterContestParticipant.ts
@@ -1,7 +1,7 @@
 import { useMutation } from "@tanstack/react-query";
 import { useNavigate } from "@tanstack/react-router";
+import { RegisterContestParticipantFunctionResponse } from "@vtuber-stamp-rally/shared-lib/functions/registerContestParticipant.ts";
 import { useTranslation } from "react-i18next";
-import { RegisterContestParticipantFunctionResponse } from "shared-lib";
 
 import { functions } from "@/lib/appwrite.ts";
 import { registerContestParticipantFunctionId } from "@/lib/consts.ts";
