+++ web/src/lib/hooks/useStandists.ts
@@ -1,7 +1,11 @@
 import { captureException } from "@sentry/react";
 import { useQuery } from "@tanstack/react-query";
+import { importJWK } from "@vtuber-stamp-rally/shared-lib/crypto.ts";
+import {
+  Standist,
+  StandistDocument,
+} from "@vtuber-stamp-rally/shared-lib/models/Standist.ts";
 import { Polygon } from "geojson";
-import { Standist, StandistDocument, importJWK } from "shared-lib";
 
 import { QUERY_KEYS } from "@/lib/QueryKeys.ts";
 import { databases } from "@/lib/appwrite.ts";
