+++ web/src/lib/images.ts
@@ -1,5 +1,5 @@
 import * as Sentry from "@sentry/react";
-import { type Standist } from "shared-lib";
+import { type Standist } from "@vtuber-stamp-rally/shared-lib/models/Standist.ts";
 
 import fallback from "@/assets/fallback-avatar.png";
