+++ web/src/routes/_rallyists/stamp.tsx
@@ -1,9 +1,9 @@
 import { useConfetti } from "@stevent-team/react-party";
 import { createFileRoute } from "@tanstack/react-router";
+import { StampTupleSerializer } from "@vtuber-stamp-rally/shared-lib/models/Stamp.ts";
 import { ListChecks, TicketCheck } from "lucide-react";
 import { LegacyRef, Suspense, useEffect } from "react";
 import { useTranslation } from "react-i18next";
-import { StampTupleSerializer } from "shared-lib";
 
 import Intro from "@/components/Intro.tsx";
 import { ArtistImage } from "@/components/artists/ArtistImage";
